gmgn_monitor/requirements.txt
websockets==12.0
python-dotenv>=1.0.0
prometheus-client>=0.19.0
colorama>=0.4.6
